herdando de classe abstrata para dao
